requirements-test.txt
#-r requirements.txt
django-nose==1.4
factory-boy==2.5.2
psycopg2==2.4.5

flake8==2.3.0
coverage==3.7.1


#mock>=1.0.1
#nose>=1.3.0
#
#httpretty
#sure>=1.2.5
#coveralls

